requirements/base.txt
Django==1.10.7
django-dotenv==1.4.1
# http://docs.wagtail.io/en/v1.8.1/topics/search/backends.html#elasticsearch-backend
# Not utilized by default; uncomment and review the above document if you require elasticsearch
# elasticsearch==5.1.0
wagtail==1.9.1
wagtailfontawesome==1.0.6
Pillow==4.0.0
google-cloud==0.25.0